requirements.txt
PySide6
opencv-python
numpy
numba
Pillow
ffmpeg-python
